another try

+++ tests/tt_metal/tt_metal/perf_microbenchmark/dispatch/common.h
@@ -45,8 +45,8 @@ private:
     bool banked;  // TODO banked and unbanked tests still don't play nicely together
     int amt_written;
     // 10 is a hack...bigger than any core_type
-    uint64_t base_data_addr[CoreType::WORKER + 1];
-    uint64_t base_result_data_addr[CoreType::WORKER + 1];
+    uint64_t base_data_addr[static_cast<int>(CoreType::WORKER) + 1];
+    uint64_t base_result_data_addr[static_cast<int>(CoreType::WORKER) + 1];
     std::unordered_map<CoreCoord, std::unordered_map<uint32_t, one_core_data_t>> all_data;
     CoreCoord host_core;
